+++ server-api/src/functional-api/activity-logs/challenge-activity-logs.it-spec.ts
@@ -26,8 +26,8 @@ import { createWhiteboardOnCallout } from '../callout/call-for-whiteboards/white
 import { assignRoleToUser, joinRoleSet } from '../roleset/roles-request.params';
 import { deleteOrganization } from '../contributor-management/organization/organization.request.params';
 import { UniqueIDGenerator } from '@alkemio/tests-lib';
-import { OrganizationWithSpaceModelFactory } from '@utils/contexts/OrganizationWithSpaceFactory';
-import { OrganizationWithSpaceModel } from '@utils/contexts/types/OrganizationWithSpaceModel';
+import { OrganizationWithSpaceModelFactory } from '@src/models/OrganizationWithSpaceFactory';
+import { OrganizationWithSpaceModel } from '@src/models/types/OrganizationWithSpaceModel';
 const uniqueId = UniqueIDGenerator.getID();
 
 let calloutDisplayName = '';
