+++ server-api/src/functional-api/journey/subspace/create-subspace.it-spec.ts
@@ -7,8 +7,8 @@ import {
   getSubspacesData,
 } from './subspace.request.params';
 import { UniqueIDGenerator } from '@alkemio/tests-lib';
-import { OrganizationWithSpaceModelFactory } from '@utils/contexts/OrganizationWithSpaceFactory';
-import { OrganizationWithSpaceModel } from '@utils/contexts/types/OrganizationWithSpaceModel';
+import { OrganizationWithSpaceModelFactory } from '@src/models/OrganizationWithSpaceFactory';
+import { OrganizationWithSpaceModel } from '@src/models/types/OrganizationWithSpaceModel';
 ;
 const uniqueId = UniqueIDGenerator.getID();
