+++ server-api/src/functional-api/journey/subspace/query-subspace-data.it-spec.ts
@@ -13,8 +13,8 @@ import {
   updateSpaceContext,
 } from '../space/space.request.params';
 import { UniqueIDGenerator } from '@alkemio/tests-lib';
-import { OrganizationWithSpaceModelFactory } from '@utils/contexts/OrganizationWithSpaceFactory';
-import { OrganizationWithSpaceModel } from '@utils/contexts/types/OrganizationWithSpaceModel';
+import { OrganizationWithSpaceModelFactory } from '@src/models/OrganizationWithSpaceFactory';
+import { OrganizationWithSpaceModel } from '@src/models/types/OrganizationWithSpaceModel';
 
 const uniqueId = UniqueIDGenerator.getID();
