+++ server-api/src/functional-api/journey/subsubspace/subsubspace_authorization.it-spec.ts
@@ -11,8 +11,8 @@ import {
   removeRoleFromUserExtendedData,
 } from '../../roleset/roles-request.params';
 import { UniqueIDGenerator } from '@alkemio/tests-lib';
-import { OrganizationWithSpaceModelFactory } from '@utils/contexts/OrganizationWithSpaceFactory';
-import { OrganizationWithSpaceModel } from '@utils/contexts/types/OrganizationWithSpaceModel';
+import { OrganizationWithSpaceModelFactory } from '@src/models/OrganizationWithSpaceFactory';
+import { OrganizationWithSpaceModel } from '@src/models/types/OrganizationWithSpaceModel';
 ;
 const uniqueId = UniqueIDGenerator.getID();
