+++ server-api/src/models/types/OrganizationWithSpaceModel.ts
